Task 5: Checking availability of a phone number

+++ include/analysis.h
@@ -8,6 +8,6 @@ unsigned short country_with_most_records(const phone_record_t *records, unsigned
 bool phone_number_exists(const phone_record_t *records,
                          unsigned long long count,
                          unsigned short country_code,
-                         unsigned int phone_number);
+                         unsigned long long phone_number);
 
 #endif
